# [POAE7-2863] optimize countSetBits and allocBuffer, and close march=native which cause coredump in double*double test