--- a/cpp/src/cider/exec/module/batch/CiderArrowBufferHolder.cpp
+++ b/cpp/src/cider/exec/module/batch/CiderArrowBufferHolder.cpp
@@ -63,7 +63,6 @@ void CiderArrowArrayBufferHolder::allocBuffer(size_t index, size_t bytes) {
     }
   } else {
     buffers_[index] = allocator_->allocate(bytes);
-    memset(buffers_[index], 0, bytes);  // set new initialized bytes to 0
     buffers_bytes_[index] = bytes;
   }
 }