--- a/cpp/src/cider/exec/nextgen/context/CodegenContext.h
+++ b/cpp/src/cider/exec/nextgen/context/CodegenContext.h
@@ -72,7 +72,7 @@ struct CodegenOptions {
   bool check_bit_vector_clear_opt = false;
   bool set_null_bit_vector_opt = false;
   bool branchless_logic = true;
-  bool enable_vectorize = false;
+  bool enable_vectorize = true;
 
   jitlib::CompilationOptions co = jitlib::CompilationOptions{};
 };