--- a/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITEngine.h
+++ b/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITEngine.h
@@ -31,9 +31,9 @@ struct CompilationOptions {
   bool optimize_ir = true;
   bool aggressive_jit_compile = true;
   bool dump_ir = false;
-  bool enable_vectorize = false;
+  bool enable_vectorize = true;
   bool enable_avx2 = true;
-  bool enable_avx512 = false;
+  bool enable_avx512 = true;
 };
 
 struct LLVMJITEngine {