--- a/cpp/src/cider/util/CiderBitUtils.h
+++ b/cpp/src/cider/util/CiderBitUtils.h
@@ -185,6 +185,7 @@ FORCE_INLINE void setBitAtUnified(uint8_t* bit_vector, size_t index, bool is_nul
 
 ENABLE_AVX256 inline size_t countSetBits(const uint8_t* bit_vector, size_t end) {
   size_t ans = 0;
+
   // group by 64
   size_t upper = end / 64;
   for (size_t i = 0; i < upper; ++i) {
@@ -192,15 +193,14 @@ ENABLE_AVX256 inline size_t countSetBits(const uint8_t* bit_vector, size_t end) 
   }
 
   // group by 8
-  size_t curr = upper * 64;
-  const uint8_t* tmp_vector = &bit_vector[upper * 64];
-  upper = (end - curr) / 8;
-  for (size_t i = 0; i < upper; ++i) {
-    ans += __builtin_popcount(tmp_vector[i]);
+  size_t curr = upper * 8;  // byte index
+  upper = end / 8;
+  for (size_t i = curr; i < upper; ++i) {
+    ans += __builtin_popcount(bit_vector[i]);
   }
 
   // remaining
-  curr += upper * 8;
+  curr = upper * 8;  // bit index
   for (size_t i = curr; i < end; ++i) {
     ans += isBitSetAt(bit_vector, i);
   }